removing the unused function
