+++ apiserver/plane/app/views/exporter/base.py
@@ -1,138 +1,25 @@
 import logging
 
-# Python imports
-from django.http import HttpResponse
-from django.utils import timezone
-
 # Third Party imports
 from rest_framework import status
 from rest_framework.response import Response
 
 from plane.app.permissions import allow_permission, ROLE
 from plane.app.serializers import ExporterHistorySerializer
-from plane.bgtasks.export_task import (
-    build_export_queryset,
-    create_zip_file,
-    generate_csv,
-    generate_json,
-    generate_xlsx,
-    issue_export_task,
-    resolve_export_columns,
-)
-from django.db.models import Exists, OuterRef, Q
+from plane.bgtasks.export_task import issue_export_task
 
 from plane.db.models import (
     ExporterHistory,
-    Issue,
     Project,
-    ProjectMember,
-    State,
     Workspace,
 )
 from plane.settings.storage import S3Storage
-from plane.utils.issue_filters import (
-    apply_user_hub_filters,
-    build_custom_property_q_objects,
-    issue_filters,
-)
-from plane.utils.order_queryset import order_issue_queryset
+from plane.utils.issue_filters import issue_filters
 
 # Module imports
 from .. import BaseAPIView
 
 logger = logging.getLogger("plane.exporter")
-
-# Hard cap on issue count for the synchronous DownloadIssuesEndpoint. Above
-# this, force the user onto the async export path so we don't OOM a request
-# worker on the cartesian explosion produced by `.values(...)` over
-# assignees + labels.
-DOWNLOAD_ISSUE_LIMIT = 10_000
-
-
-_LEGACY_ISSUE_EXPORT_HEADER = [
-    "ID",
-    "Project",
-    "Name",
-    "Description",
-    "State",
-    "Priority",
-    "Created By",
-    "Assignee",
-    "Labels",
-    "Cycle Name",
-    "Cycle Start Date",
-    "Cycle End Date",
-    "Module Name",
-    "Module Start Date",
-    "Module Target Date",
-    "Created At",
-    "Updated At",
-    "Completed At",
-    "Archived At",
-]
-
-EXPORTER_MAPPER = {
-    "csv": generate_csv,
-    "json": generate_json,
-    "xlsx": generate_xlsx,
-}
-
-def _build_list_view_base_queryset(slug, user, project_ids):
-    """Mirror the baseline queryset used by WorkspaceViewIssuesViewSet.list so
-    the export contains exactly the rows the user sees on the page (drafts /
-    archived / triage states excluded, hub-scoped, guest restrictions applied).
-    """
-    qs = (
-        Issue.objects.filter(
-            workspace__slug=slug,
-            project_id__in=project_ids,
-            deleted_at__isnull=True,
-            archived_at__isnull=True,
-            is_draft=False,
-            project__archived_at__isnull=True,
-            state_id__isnull=False,
-        )
-        .exclude(
-            state_id__in=State.objects.filter(is_triage=True).values("id")
-        )
-        .filter(
-            Exists(
-                ProjectMember.objects.filter(
-                    project_id=OuterRef("project_id"),
-                    member=user,
-                    is_active=True,
-                )
-            )
-        )
-    )
-
-    # Mirror the guest_view_all_features split applied by the list view.
-    guest_pm = ProjectMember.objects.filter(
-        project_id=OuterRef("project_id"),
-        member=user,
-        is_active=True,
-        role=5,
-    )
-    non_guest_pm = ProjectMember.objects.filter(
-        project_id=OuterRef("project_id"),
-        member=user,
-        is_active=True,
-        role__gt=5,
-    )
-    qs = qs.filter(
-        Exists(non_guest_pm)
-        | (Exists(guest_pm) & Q(project__guest_view_all_features=True))
-        | (Exists(guest_pm) & Q(project__guest_view_all_features=False) & Q(created_by=user))
-    )
-
-    return apply_user_hub_filters(qs, user, workspace_slug=slug)
-
-
-CONTENT_TYPES = {
-    "csv": "text/csv",
-    "json": "application/json",
-    "xlsx": "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
-}
 
 
 class ExportIssuesEndpoint(BaseAPIView):
@@ -148,7 +35,7 @@ class ExportIssuesEndpoint(BaseAPIView):
         multiple = request.data.get("multiple", False)
         project_ids = request.data.get("project", [])
 
-        if provider in ["csv", "xlsx", "json"]:
+        if provider == "csv":
             if not project_ids:
                 project_ids = Project.objects.filter(
                     workspace__slug=slug,
@@ -256,117 +143,3 @@ class ExportIssuesEndpoint(BaseAPIView):
                 {"error": "per_page and cursor are required"},
                 status=status.HTTP_400_BAD_REQUEST,
             )
-
-
-class DownloadIssuesEndpoint(BaseAPIView):
-    """Synchronously build and return the export file as a download. No Celery, no S3."""
-
-    @allow_permission(allowed_roles=[ROLE.ADMIN, ROLE.MEMBER], level="WORKSPACE")
-    def post(self, request, slug):
-        provider = request.data.get("provider", False)
-        multiple = request.data.get("multiple", False)
-        project_ids = request.data.get("project", [])
-
-        if provider not in ["csv", "xlsx", "json"]:
-            return Response(
-                {"error": f"Provider '{provider}' not found."},
-                status=status.HTTP_400_BAD_REQUEST,
-            )
-
-        workspace = Workspace.objects.get(slug=slug)
-
-        if not project_ids:
-            project_ids = list(
-                Project.objects.filter(
-                    workspace__slug=slug,
-                    project_projectmember__member=request.user,
-                    project_projectmember__is_active=True,
-                    archived_at__isnull=True,
-                ).values_list("id", flat=True)
-            )
-            project_ids = [str(project_id) for project_id in project_ids]
-
-        # Parse same filters the issues list view uses (assignees, state,
-        # labels, priority, sub_issue, cycle, module, start_date, target_date,
-        # search/name, hub/customer/vendor/worker/business fields,
-        # custom_properties, etc.) from the request query string so the export
-        # matches what the user sees on the frontend.
-        filters = issue_filters(request.query_params, "GET")
-        custom_properties = filters.pop("custom_properties", {}) or {}
-        custom_filters = build_custom_property_q_objects(custom_properties)
-        order_by_param = request.query_params.get("order_by", "-created_at")
-        display_properties_raw = request.query_params.get("display_properties")
-        display_properties = (
-            [k.strip() for k in display_properties_raw.split(",") if k.strip()]
-            if display_properties_raw
-            else None
-        )
-        columns = resolve_export_columns(display_properties)
-        header = [r.label for r in columns]
-
-        base_qs = (
-            _build_list_view_base_queryset(
-                slug=slug,
-                user=request.user,
-                project_ids=project_ids,
-            )
-            .filter(*custom_filters)
-            .filter(**filters)
-        )
-
-        # Guard: hard cap on issue count. Count distinct Issue IDs *before*
-        # the cartesian `.values(...)` join with assignees/labels — counting
-        # after that explodes memory. If the user is over the limit, point
-        # them at the async export path which writes to S3 and doesn't tie
-        # up a request worker.
-        issue_count = base_qs.values("id").distinct().count()
-        if issue_count > DOWNLOAD_ISSUE_LIMIT:
-            return Response(
-                {
-                    "error": (
-                        f"This download would include {issue_count} issues, "
-                        f"which exceeds the {DOWNLOAD_ISSUE_LIMIT} limit. "
-                        "Use the async export instead, or narrow your filters."
-                    ),
-                    "issue_count": issue_count,
-                    "limit": DOWNLOAD_ISSUE_LIMIT,
-                },
-                status=status.HTTP_400_BAD_REQUEST,
-            )
-
-        base_qs, _ = order_issue_queryset(
-            issue_queryset=base_qs,
-            order_by_param=order_by_param,
-        )
-        # Build the final values()-queryset from the resolved columns so
-        # this endpoint and issue_export_task emit identical rows for the
-        # same filters (including custom/introspected Issue fields).
-        workspace_issues = build_export_queryset(base_qs, columns)
-
-        exporter = EXPORTER_MAPPER.get(provider)
-        files = []
-
-        if multiple:
-            for project_id in project_ids:
-                issues = workspace_issues.filter(project__id=project_id)
-                exporter(header, project_id, issues, files, columns)
-        else:
-            exporter(header, str(workspace.id), workspace_issues, files, columns)
-
-        date_str = timezone.now().date().isoformat()
-
-        # Single project / single file → send the raw file. Multi-file → zip it.
-        if len(files) == 1 and not multiple:
-            filename, content = files[0]
-            download_name = f"{slug}-issues-{date_str}.{provider}"
-            content_type = CONTENT_TYPES[provider]
-            response_body = content if isinstance(content, (bytes, bytearray)) else content.encode("utf-8")
-            response = HttpResponse(response_body, content_type=content_type)
-            response["Content-Disposition"] = f'attachment; filename="{download_name}"'
-            return response
-
-        zip_buffer = create_zip_file(files)
-        download_name = f"{slug}-issues-{date_str}.zip"
-        response = HttpResponse(zip_buffer.getvalue(), content_type="application/zip")
-        response["Content-Disposition"] = f'attachment; filename="{download_name}"'
-        return response
